move path stuff into its own module
